Final Version Merge

+++ Controllers/VenueController.cs
@@ -2,6 +2,9 @@ using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Linq;
+using System.Numerics;
+using System.Text.Json;
+using System.Text.Json.Serialization;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
@@ -9,9 +12,10 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using WebCoursework.Models;
 
+
 namespace WebCoursework.Controllers
 {
-    [Authorize]
+    [Authorize(Roles = "Admin,User")]
     [Route("api/[controller]")]
     [ApiController]
     public class VenueController : ControllerBase
@@ -29,6 +33,7 @@ namespace WebCoursework.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<Venue>>> GetVenue()
         {
+            _logger.LogInformation("Venues successfully retrieved");
             return await _context.Venue.ToListAsync();
         }
 
@@ -40,17 +45,33 @@ namespace WebCoursework.Controllers
 
             if (venue == null)
             {
-                return NotFound();
+                return NotFound("Venue not found");
             }
-
+            _logger.LogInformation($"Venue (ID: {id}) successfully retrieved");
             return venue;
         }
 
         // PUT: api/VenueContext/5
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
         [HttpPut("{id}")]
         public async Task<IActionResult> PutVenue(int id, Venue venue)
         {
+            if (!VenueExists(id))
+            {
+                return VenueNotExistMessage(id);
+            }
+
+            if (id != venue.VenueId)
+            {
+                _logger.LogInformation($"URL ID {id} doesn't match request for Venue ID {venue.VenueId}");
+                return BadRequest($"The Venue ID in the URL ({id}) does not match the Venue ID ({venue.VenueId}) in the request body");
+            }
+
+            if (venue.Capacity > 100000 || venue.Capacity < 10000)
+            {
+                _logger.LogInformation("An attempt to change a venue's capacity beyond league rules was made.");
+                return BadRequest("League rules state you cannot increase a venue's capacity to over 100,000 or below 10,000.");
+            }
+
             if (id != venue.VenueId)
             {
                 return BadRequest();
@@ -73,40 +94,65 @@ namespace WebCoursework.Controllers
                     throw;
                 }
             }
-
+            _logger.LogInformation($"Venue (ID: {venue.VenueId}) successfully edited");
             return NoContent();
         }
 
         // POST: api/VenueContext
-        // To protect from overposting attacks, see https://go.microsoft.com/fwlink/?linkid=2123754
+        [Authorize(Roles = "Admin")]
         [HttpPost]
         public async Task<ActionResult<Venue>> PostVenue(Venue venue)
         {
+            if (venue.Capacity > 100000 || venue.Capacity < 10000)
+            {
+                _logger.LogInformation("An attempt to create a venue with capacity beyond league rules was made.");
+                return BadRequest("League rules state you cannot have a venue's capacity over 100,000 or below 10,000.");
+            }
+
+
             _context.Venue.Add(venue);
             await _context.SaveChangesAsync();
-
+            _logger.LogInformation($"Venue (ID: {venue.VenueId}) successfully created");
             return CreatedAtAction("GetVenue", new { id = venue.VenueId }, venue);
         }
 
         // DELETE: api/VenueContext/5
+        [Authorize(Roles = "Admin")]
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteVenue(int id)
         {
+            if (!VenueExists(id))
+            {
+                return VenueNotExistMessage(id);
+            }
+
             var venue = await _context.Venue.FindAsync(id);
             if (venue == null)
             {
                 return NotFound();
             }
 
+            if (_context.Match.Any(m => m.VenueId == id))
+            {
+                _logger.LogInformation($"Failed to delete a venue, ID: ({id}) as holds a record of a match played here");
+                return BadRequest($"Venue ID: {id} has hosted matches so can't be deleted.");
+            }
+
             _context.Venue.Remove(venue);
             await _context.SaveChangesAsync();
-
+            _logger.LogInformation($"Venue (ID: {id}) successfully deleted");
             return NoContent();
         }
 
         private bool VenueExists(int id)
         {
             return _context.Venue.Any(e => e.VenueId == id);
+        }
+
+        private IActionResult VenueNotExistMessage(int id)
+        {
+            _logger.LogInformation($"Failed to find a Venue with Id ({id}) passed by the user");
+            return BadRequest($"A Venue with ID {id} does not exist");
         }
     }
 }
